docs(catalog): embed catalog UI in docs via iframe + post-build hook
- Rename copy_changelog.py → pre_post_build.py (broader scope)
- Add on_post_build hook that generates catalog.html in site/ from
  catalog_wrapper.html template + shared catalog fragment
- catalog_wrapper.html: standalone Bootstrap page with system theme
  detection, toggle button, and navbar
- Docs catalog page embeds via iframe pointing to /catalog.html
- Move filter input to the right, next to plugin count
- Disclaimer displayed above the filter row
- Add .DS_Store to .gitignore
- Update CHANGELOG.md
