+++ docs/plugins/catalog.md
@@ -2,6 +2,7 @@
 
 The following plugins are available for az-scout. Plugins can be installed directly from the Plugin Manager in your az-scout instance, or manually using pip/uv.
 
---8<--
-docs/_includes/catalog.html
---8<--
+<iframe id="catalog-frame" src="/catalog.html" style="width:100%;min-height:800px;border:none;" loading="lazy"></iframe>
+
+!!! info "Local preview"
+    The embedded catalog requires a built site (`mkdocs build`). Use `python -m http.server -d site` to preview locally.
