+++ tools/docs_hooks/catalog_wrapper.html
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <title>Azure Scout – Plugin Catalog</title>
+  <link rel="icon" href="https://docs.az-scout.com/assets/favicon.svg">
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.6/dist/css/bootstrap.min.css"
+        rel="stylesheet" crossorigin="anonymous">
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
+        rel="stylesheet" crossorigin="anonymous">
+  <style>
+    body { padding: 1rem; }
+    .catalog-theme-toggle { position: fixed; top: 0.5rem; right: 0.5rem; z-index: 1000; }
+    .navbar { margin-bottom: 1rem; }
+  </style>
+  <script>
+    (function() {
+      var stored = localStorage.getItem("catalog-theme");
+      var theme = stored ||
+        (window.matchMedia("(prefers-color-scheme: dark)").matches ? "dark" : "light");
+      document.documentElement.setAttribute("data-bs-theme", theme);
+    })();
+  </script>
+</head>
+<body>
+
+<!-- Header -->
+<nav class="navbar border-bottom">
+    <div class="container-fluid px-4">
+        <span class="navbar-brand mb-0 h1">Azure Scout <span class="text-body-secondary fw-normal">– Plugin
+                Catalog</span></span>
+        <div class="d-flex align-items-center gap-2">
+            <button class="btn btn-sm btn-outline-secondary" id="themeToggle" title="Toggle theme">🌙</button>
+        </div>
+    </div>
+</nav>
+
+<!-- CATALOG_FRAGMENT -->
+
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.6/dist/js/bootstrap.bundle.min.js"
+        crossorigin="anonymous"></script>
+<script>
+  var toggle = document.getElementById("themeToggle");
+  function updateToggle() {
+    toggle.textContent = document.documentElement.getAttribute("data-bs-theme")
+      === "dark" ? "\u2600\uFE0F" : "\uD83C\uDF19";
+  }
+  updateToggle();
+  toggle.addEventListener("click", function() {
+    var next = document.documentElement.getAttribute("data-bs-theme")
+      === "dark" ? "light" : "dark";
+    document.documentElement.setAttribute("data-bs-theme", next);
+    localStorage.setItem("catalog-theme", next);
+    updateToggle();
+  });
+</script>
+</body>
+</html>
